+++ rest/src/decorators/index.ts
@@ -1,0 +1,2 @@
+export * from './handle404.decorator';
+export * from './ignoreClientErrors.decorator';