+++ rest/src/errorHandlers/index.ts
@@ -1,0 +1,2 @@
+export * from './handle400WithValidationsFunc';
+export * from './handle404Func';