+++ rest/src/index.ts
@@ -1,0 +1,9 @@
+export * from './decorators';
+export * from './errorHandlers';
+export * from './interceptors';
+export * from './middlewares';
+export * from './rxjsOperators';
+
+export * from './rest.interface';
+export * from './tokens';
+export * from './types';