+++ rest/src/rxjsOperators/handle404.operator.ts
@@ -1,0 +1,17 @@
+import {HttpErrorResponse} from '@angular/common/http';
+import {MonoTypeOperatorFunction} from 'rxjs';
+import {catchError} from 'rxjs/operators';
+
+import {handle404Func} from '../errorHandlers';
+import {RestNotFoundError} from '../types';
+
+/**
+ * Handles 404 http code as response
+ */
+export function handle404(): MonoTypeOperatorFunction<RestNotFoundError|HttpErrorResponse>
+{
+    return source =>
+    {
+        return source.pipe(catchError(handle404Func));
+    };
+}