+++ rest/src/rxjsOperators/index.ts
@@ -1,0 +1,2 @@
+export * from './handle400WithValidations.operator';
+export * from './handle404.operator';